+++ src/MF_Servo/Servos.cpp
@@ -13,7 +13,7 @@ namespace Servos
     MFServo *servos[MAX_MFSERVOS];
     uint8_t  servosRegistered = 0;
 
-    void     Add(int pin)
+    void     Add(uint8_t pin)
     {
         if (servosRegistered == MAX_MFSERVOS)
             return;
